moving debugborder to helper, remove navigation view
