+++ Sources/TMDB_MVVM_Detail/Views/Sections/MovieCreditSection.swift
@@ -30,27 +30,6 @@ struct MovieCreditSection: View {
     }
 }
 
-extension Color {
-    static func random() -> Color {
-        return Color(
-            red: Double.random(in: 0 ... 1),
-            green: Double.random(in: 0 ... 1),
-            blue: Double.random(in: 0 ... 1)
-        )
-    }
-}
-
-extension View {
-    @ViewBuilder
-    func debugBorder(color: Color? = nil) -> some View {
-#if DEBUG
-        border(color ?? Color.random(), width: 0.5)
-#else
-        self
-#endif
-    }
-}
-
 struct OnFirstAppearModifier: ViewModifier {
     let perform: () -> Void
     @State private var firstTime: Bool = true
